+++ brainpy/connectivity/__init__.py
@@ -2,5 +2,5 @@
 
 
 from .base import *
-from .formater import *
+from .formatter import *
 from .methods import *