+++ brainpy/integration/__init__.py
@@ -3,6 +3,7 @@
 from . import diff_equation
 from . import integrator
 from . import sympy_tools
+from . import methods
 from .diff_equation import *
 from .integrator import *
 from .. import profile