+++ brainpy/profile.py
@@ -13,8 +13,15 @@ Using the API in ``profile.py``, you can set
 """
 
 __all__ = [
+    'set',
     'set_backend',
     'get_backend',
+
+    'is_jax_bk',
+    'is_numpy_bk',
+    'is_numba_bk',
+    'set_numba_profile',
+    'get_numba_profile',
 
     'set_dt',
     'get_dt',